refactor(tests): Split removeSubtask tests
